Move from_script to Address and drop AddressExt
With the prior changes to AddressExt::from_script, the method can now
be directly moved to the main Address type. With said move, the
extension trait can also be dropped entirely.

Move AddressExt::from_script to the Address type in addresses.
Drop the AddressExt trait.
